tvempresa/requirements/base.txt

python-dotenv==0.7.1
psycopg2==2.8.3
Pillow==6.1.0
boto3==1.9.202
pycurl==7.43.0.3
djangorestframework==3.10.2
beautifulsoup4==4.8.0
django-uuslug==1.1.8
django-import-export==1.2.0
django-admin-rangefilter==0.5.0
django-ckeditor==5.7.1
singlemodeladmin==0.9
django-admin-row-actions==0.0.5
django-pure-pagination==0.3.0
django-zipview==1.2.0
requests==2.22.0
django-cookie-law==2.0.2
python-nginx==1.5.3
django-taggit==1.1.0
django-filter==2.2.0
django==2.2.4
sorl-thumbnail==12.7.0